Masaüstü/trade/requirements.txt
python-binance>=1.0.19
requests>=2.31.0
pillow>=10.0.0
pynacl>=1.5.0
